requirement.txt
Django
djangorestframework
django-rest-auth
django-cors-headers
psycopg2
gunicorn
phonenumbers
django-phonenumber-field